authorization express

+++ controllers/cards.js
@@ -1,15 +1,16 @@
 const Card = require('../models/card');
 
-const ERROR_VALIDATION = 400;
-const ERROR_NOT_FOUND = 404;
+const NotFoundError = require('../errors/NotFoundError');
+const ValidationError = require('../errors/ValidationError');
+const ForbiddenError = require('../errors/ForbiddenError');
 
-const getCards = (req, res) => {
+const getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.status(200).send(cards))
-    .catch((err) => res.status(500).send(err));
+    .catch(next);
 };
 
-const createCard = (req, res) => {
+const createCard = (req, res, next) => {
   const { name, link } = req.body;
   const owner = req.user._id;
 
@@ -17,47 +18,65 @@ const createCard = (req, res) => {
     .then((card) => res.send(card))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(ERROR_VALIDATION).send({ message: 'Данные не прошли валидацию' });
+        next(new ValidationError('Указаны некорректные данные'));
+        return;
       }
-      res.status(500).send(err);
+      next(err);
     });
 };
 
-const deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+const deleteCard = (req, res, next) => {
+  const owner = req.user._id;
+  Card.findById(req.params.cardId)
     .then((card) => {
-      if (!card) {
-        res.status(ERROR_NOT_FOUND).send({ message: 'Запрашиваемая карточка не найдена' });
+      if (owner === card.owner) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then(() => {
+            if (!card) {
+              next(new NotFoundError('Запрашиваемая карточка не найдена'));
+              return;
+            }
+            res.status(200).send('Карточка удалена');
+          })
+          .catch((err) => {
+            if (err.name === 'CastError') {
+              next(new ValidationError('Невалидный id карточки'));
+              return;
+            }
+            res.status(500).send(err);
+          });
       }
-      res.status(200).send('Карточка удалена');
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(ERROR_VALIDATION).send({ message: 'Невалидный id карточки' });
+        next(new ForbiddenError('Вы не можете удалить эту карточку'));
+        return;
       }
-      res.status(500).send(err);
+      next(err);
     });
 };
 
-const putLike = (req, res) => {
+const putLike = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId,
     { $addToSet: { likes: req.user._id } },
     { new: true })
     .then((card) => {
       if (!card) {
-        res.status(ERROR_NOT_FOUND).send({ message: 'Нет карточки по заданному id' });
+        next(new NotFoundError('Нет карточки по заданному id'));
+        return;
       }
       res.status(200).send(card);
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(ERROR_VALIDATION).send({ message: 'Невалидный id карточки' });
+        next(new NotFoundError('Невалидный id карточки'));
+        return;
       }
-      res.status(500).send(err);
+      next(err);
     });
 };
 
-const deleteLike = (req, res) => {
+const deleteLike = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
@@ -65,15 +84,17 @@ const deleteLike = (req, res) => {
   )
     .then((card) => {
       if (!card) {
-        res.status(ERROR_NOT_FOUND).send({ message: 'Нет карточки по заданному id' });
+        next(new NotFoundError('Нет карточки по заданному id'));
+        return;
       }
       res.status(200).send(card);
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(ERROR_VALIDATION).send({ message: 'Невалидный id карточки' });
+        next(new NotFoundError('Невалидный id карточки'));
+        return;
       }
-      res.status(500).send(err);
+      next(err);
     });
 };
